requirements.txt
appdirs==1.4.4
Babel==2.9.1
black==21.7b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-WTF==0.15.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
libsass==0.21.0
MarkupSafe==2.0.1
mongoengine==0.24.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
pydantic==1.8.2
pymongo==3.12.0
pytz==2021.1
regex==2021.8.3
requests==2.26.0
six==1.16.0
tomli==1.2.1
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
gunicorn==20.1.0